Start implementing contentful
Implementing contenful is not as easy as it seems with the current NextJS implementation because of the way the components are structured. It seems like I'll need to use a different method for passing in the contenful content, or restructure the components so that they work with the content that comes in.